Move command factory to separate class
Register subclasses in command factory
